Create StatusBar component with safeAreaView
