--- front-end-mobile/src/features/loading/LoadingScreen.styles.ts
@@ -1,16 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-const styles = StyleSheet.create({
-  container: { 
-    flex: 1, 
-    flexDirection: 'column', 
-    justifyContent: 'center', 
-    alignItems: 'center',
-    backgroundColor: 'white', 
-  },
-  dotLoaders: {
-    marginTop: 20,
-  },
-});
-
-export default styles;
