+++ front-end-mobile/src/features/loading/LoadingScreen.tsx
@@ -1,19 +1,32 @@
 import React from 'react';
-import { SafeAreaView, Image } from 'react-native';
+import { Image, StyleSheet, View } from 'react-native';
 import R from 'res/R';
 import DotsLoader from './components/DotsLoader';
-import styles from './LoadingScreen.styles';
+import StatusBar from 'res/components/statusBar/StatusBar';
 
 const LoadingScreen = () => {
   return (
     <>
-      <SafeAreaView
-        style={styles.container}>
+      <StatusBar backgroundColor={R.colors.WHITE} barStyle='dark-content' />
+      <View style={styles.container}>
         <Image source={R.images.ic_splash} />
         <DotsLoader style={styles.dotLoaders} width={100} dotNumber={4} />
-      </SafeAreaView>
+      </View>
     </>
   );
 };
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    flexDirection: 'column',
+    justifyContent: 'center',
+    alignItems: 'center',
+    backgroundColor: R.colors.WHITE,
+  },
+  dotLoaders: {
+    marginTop: 20,
+  },
+});
 
 export default LoadingScreen;
